Conflicts:
	Installer/VersionInfo.ini
	MultiMiner.CoinApi/Properties/AssemblyInfo.cs
	MultiMiner.CoinChoose/Properties/AssemblyInfo.cs
	MultiMiner.CoinWarz/Properties/AssemblyInfo.cs
	MultiMiner.Coinbase/Properties/AssemblyInfo.cs
	MultiMiner.Discovery/Properties/AssemblyInfo.cs
	MultiMiner.Engine/Properties/AssemblyInfo.cs
	MultiMiner.MobileMiner/Properties/AssemblyInfo.cs
	MultiMiner.Remoting/Properties/AssemblyInfo.cs
	MultiMiner.Services/Properties/AssemblyInfo.cs
	MultiMiner.Stats/Properties/AssemblyInfo.cs
	MultiMiner.Update/Properties/AssemblyInfo.cs
	MultiMiner.Utility/Properties/AssemblyInfo.cs
	MultiMiner.Win/Properties/AssemblyInfo.cs
	MultiMiner.Xgminer.Api/Properties/AssemblyInfo.cs
	MultiMiner.Xgminer.Discovery/Properties/AssemblyInfo.cs
	MultiMiner.Xgminer/Properties/AssemblyInfo.cs

+++ MultiMiner.Win/ViewModels/DeviceViewModel.cs
@@ -51,7 +51,7 @@ namespace MultiMiner.Win.ViewModels
         public double LastShareDifficulty { get; set; }
         public DateTime? LastShareTime { get; set; }
         public string Url { get; set; }
-        public int BestShare { get; set; }
+        public double BestShare { get; set; }
         public double PoolStalePercent { get; set; }
 
         //worker info
